Updated Channel_MODULE.cc's note_off override to force calc_fade_per_tick to true, not false. Updated convert_xm_envelope in Load_XM.cc and load_it_convert_envelope in Load_IT.cc to generate volume envelopes that are always looping. Just, if they are specified as non-looping, that just means they're looping on their final frame. Fixed Load_XM.cc to not skip all notes done in the last instrument. Fixed note fadeout: - Updated the fade_out field of struct sample to be a double. - Updated convert_samples in Load_XM.cc to convert xm_header.fade_out to a meaningful double value for fade per frame. - Updated base_note_off in channel.cc to begin fade on notes where the sample has a non-zero fade_out set. - Reworked the fade_per_tick calculation in base_note_off to only clamp to dropoff_min_length/dropoff_max_length when using dropoff_proportion, and to compute the fade duration in ticks explicitly (the dropoff-based calculation starts out with a duration in seconds).

+++ src/sample.h
@@ -53,7 +53,7 @@ namespace MultiPLAY
 		int index;
 		string name;
 
-		unsigned int fade_out;
+		double fade_out;
 
 		unsigned int num_samples;
 		double samples_per_second;
